requirements.txt
blinker==1.6.2
brotlipy==0.7.0
click==8.1.6
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
fonttools==4.42.0
fsspec==2023.6.0
greenlet==2.0.2
huggingface-hub==0.16.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpointer==2.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.8.8
safetensors==0.3.2
SQLAlchemy==2.0.19
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
transformers==4.31.0
typing_extensions==4.7.1
Werkzeug==2.3.7
